Autoroot fix.
